RunVerifier clean up
