+++ RunVerifier/package.json
@@ -19,6 +19,7 @@
     "uuid": "^8.3.2"
   },
   "devDependencies": {
-    "dotenv": "^16.0.0"
+    "dotenv": "^16.0.0",
+    "eslint": "^8.15.0"
   }
 }
